Fixing .h file

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -15,7 +15,7 @@ hash_table_t *hash_table_create(unsigned long int size)
 	if (!ht)
 		return (NULL);
 	ht->size = size;
-	ht->array = calloc((size_t)size, sizeof(hash_node_t *));
+	ht->array = calloc(size, sizeof(hash_node_t *));
 	if (!ht->array)
 		return (NULL);
 	return (ht);
